refactor: api/ 配下を Hono ベースにリファクタリング
手動 URLPattern ルーティングを Hono に移行し、ルートごとにファイル分割。
エラーハンドリングを app.onError() で一元化。

closes #32
